chore: move SearchBox to Search component
